requirements.txt
pytest
pyyaml
jsonschema
black
isort
flake8
